requirements.txt
langchain
langchain_core
langchain_google_genai
langgraph
python-dotenv
streamlit
